+++ python/aristotle-metadata-registry/aristotle_mdr/tests/utils.py
@@ -1,4 +1,5 @@
 from django import VERSION as django_version
+import attr
 import datetime
 import random
 import string
@@ -648,8 +649,19 @@ class LoggedInViewPages(object):
 
 
 class FormsetTestUtils:
+    """Utilities to help create formset post data"""
 
     def get_formset_postdata(self, datalist, prefix='form', initialforms=0):
+        """
+        Get postdata for a formset
+
+        Arguments:
+            datalist: List of form dictionaries to be posted
+
+        Keyword Arguments:
+            prefix: Formsets prefix
+            initialforms: number of forms initialy rendered
+        """
 
         postdata = {}
         # Add data
@@ -675,10 +687,11 @@ class FormsetTestUtils:
         response = self.client.post(url, postdata)
         return response
 
+
 class GeneralTestUtils:
+    """General test utilities to assist with common unit test functionality"""
 
     def _status_check(self, response, kwargs):
-
         if 'status_code' in kwargs:
             self.assertEqual(response.status_code, kwargs['status_code'])
 
@@ -696,21 +709,87 @@ class GeneralTestUtils:
         return response
 
     def reverse_get(self, *args, **kwargs):
+        """
+        Get by reverse url
+
+        Arguments:
+        url_name -- named url to reverse
+        Standard client.get args
+
+        Extra Keyword Arguments:
+        reverse_args -- args list to use during reverse
+        status_code -- expected status code of response
+        """
         return self._reverse_request('get', *args, **kwargs)
 
-    def reverse_post(self, url_name, *args, **kwargs):
+    def reverse_post(self, *args, **kwargs):
+        """
+        Post by reverse url
+
+        Arguments:
+        url_name -- named url to reverse
+        Standard client.post args
+
+        Extra Keyword Arguments:
+        reverse_args -- args list to use during reverse
+        status_code -- expected status code of response
+        """
         return self._reverse_request('post', *args, **kwargs)
 
     def assertContext(self, response, key, value):
-
+        """Check that a key and value are in context"""
         context = response.context
-        self.assertEqual(key in context)
+        self.assertTrue(key in context)
         self.assertEqual(context[key], value)
 
-    def assertInContext(response, key):
-
+    def assertInContext(self, response, key):
+        """Check that a key is in context"""
         context = response.context
-        self.assertEqual(key in context)
+        self.assertTrue(key in context)
+
+
 
 class AristotleTestUtils(LoggedInViewPages, GeneralTestUtils, FormsetTestUtils):
+    """Combination of the above 3 utils for easy usage"""
     pass
+
+
+@attr.s
+class MockManagementForm(object):
+    prefix = attr.ib(default=0)
+    max_forms = attr.ib(default=1000)
+    min_forms = attr.ib(default=0)
+    mock_form = attr.ib(default=None)
+    is_ordered = attr.ib(default=False)
+    initial_forms = attr.ib(init=False)
+    forms = []
+
+    def __attrs_post_init__(self):
+        self.forms = []
+        self.initial_forms_count = len(self.forms)
+
+    def add_forms(self, forms=[]):
+        assert type(forms) is list
+        for form in forms:
+            self.add_form(form)
+
+    def add_form(self, form={}):
+        assert type(form) is dict
+        if self.is_ordered:
+            form['ORDER'] = form.get('ORDER', len(self.forms))
+        if form:
+            self.forms.append(form)
+
+    def as_dict(self):
+        base = {
+            '{}-INITIAL_FORMS'.format(self.prefix): self.initial_forms_count,
+            '{}-TOTAL_FORMS'.format(self.prefix): len(self.forms),
+            '{}-MIN_NUM_FORMS'.format(self.prefix): self.min_forms,
+            '{}-MAX_NUM_FORMS'.format(self.prefix): self.max_forms
+        }
+        
+        for i, form in enumerate(self.forms):
+            for field, value in form.items():
+                base['{}-{}-{}'.format(self.prefix, i, field)] = value
+
+        return base